+++ src/settings/DendronNoteSettings.ts
@@ -14,10 +14,7 @@ export const _add_settings_dendron_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.dendron_note.enabled =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -32,10 +29,7 @@ export const _add_settings_dendron_note = (
 				else {
 					plugin.settings.explicit_edge_sources.dendron_note.delimiter =
 						value;
-					await Promise.all([
-						plugin.rebuildGraph(),
-						plugin.saveSettings(),
-					]);
+					await plugin.commitSettings("graph");
 				}
 			},
 		},
@@ -51,10 +45,7 @@ export const _add_settings_dendron_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.dendron_note.default_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -69,10 +60,7 @@ export const _add_settings_dendron_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.dendron_note.default_sibling_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -86,10 +74,7 @@ export const _add_settings_dendron_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.dendron_note.display_trimmed =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});