+++ src/settings/JohnnyDecimalSettings.ts
@@ -15,10 +15,7 @@ export const _add_settings_johnny_decimal_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.johnny_decimal_note.enabled =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -34,10 +31,7 @@ export const _add_settings_johnny_decimal_note = (
 				else {
 					plugin.settings.explicit_edge_sources.johnny_decimal_note.delimiter =
 						value;
-					await Promise.all([
-						plugin.rebuildGraph(),
-						plugin.saveSettings(),
-					]);
+					await plugin.commitSettings("graph");
 				}
 			},
 		},
@@ -53,10 +47,7 @@ export const _add_settings_johnny_decimal_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.johnny_decimal_note.default_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -71,10 +62,7 @@ export const _add_settings_johnny_decimal_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.johnny_decimal_note.default_sibling_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});