+++ src/settings/RebuildGraphSettings.ts
@@ -14,10 +14,7 @@ export const _add_settings_rebuild_graph = (
 			cb: async (value) => {
 				plugin.settings.commands.rebuild_graph.notify = value;
 
-				await Promise.all([
-					plugin.saveSettings(),
-					plugin.rebuildGraph(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -37,10 +34,7 @@ export const _add_settings_rebuild_graph = (
 			cb: async (value) => {
 				plugin.settings.commands.rebuild_graph.trigger = value;
 
-				await Promise.all([
-					plugin.saveSettings(),
-					plugin.rebuildGraph(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});