+++ src/settings/TagNoteSettings.ts
@@ -14,10 +14,7 @@ export const _add_settings_tag_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.tag_note.default_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});
@@ -32,10 +29,7 @@ export const _add_settings_tag_note = (
 			cb: async (value) => {
 				plugin.settings.explicit_edge_sources.tag_note.default_sibling_field =
 					value;
-				await Promise.all([
-					plugin.rebuildGraph(),
-					plugin.saveSettings(),
-				]);
+				await plugin.commitSettings("graph");
 			},
 		},
 	});